securevision_core/requirements.txt
streamlit
opencv-python
numpy
ultralytics
psycopg2-binary
streamlit-extras
